Removing pull requests when they get closed in github
Expanding the current pull request handler to deal with pull requests being
closed. Deciding on a map structure to determine which commands are run based on
each event and so removing the need to using a switch/case/if statement.

Also using the opportunity to tidy up and make sure that we use the *Repository
classes to interface with the storage classes.